+++ lib/src/report/report_sender.dart
@@ -8,21 +8,25 @@ class ReportSender {
   static Future<void> sendEmail({
     required LogReporterConfig config,
     required String report,
+    String? subject,
     BuildContext? context,
+    bool clearHistoryOnSuccess = false,
   }) async {
     final smtpServer = gmail(config.smtpEmail, config.appPassword);
     final message = Message()
       ..from = Address(config.smtpEmail, '-${config.appName}- App Logger Alert')
       ..recipients.add(config.toEmail)
-      ..subject =
+      ..subject = subject ??
           'Daily App Report :: ${config.appName}-${config.isProduction ? "PROD" : "STAGING"} :: ${DateTime.now()}'
       ..text = report;
     try {
       await send(message, smtpServer);
-      AppLogReporter.talker.cleanHistory();
-      if (context.runtimeType != Null) {
+      if (clearHistoryOnSuccess) {
+        AppLogReporter.talker.cleanHistory();
+      }
+      if (context != null) {
         WidgetsBinding.instance.addPostFrameCallback((_) {
-          ScaffoldMessenger.of(context!).showSnackBar(
+          ScaffoldMessenger.maybeOf(context)?.showSnackBar(
             SnackBar(
               content: Text('Send Successfully'),
               backgroundColor: Colors.green,
@@ -33,15 +37,17 @@ class ReportSender {
       }
     } catch (e) {
       Clipboard.setData(ClipboardData(text: e.toString()));
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        ScaffoldMessenger.of(context!).showSnackBar(
-          SnackBar(
-            content: Text('error copied , send email Erorr : ${e.toString()}'),
-            backgroundColor: Colors.red,
-            duration: Duration(seconds: 2),
-          ),
-        );
-      });
+      if (context != null) {
+        WidgetsBinding.instance.addPostFrameCallback((_) {
+          ScaffoldMessenger.maybeOf(context)?.showSnackBar(
+            SnackBar(
+              content: Text('error copied , send email Error : ${e.toString()}'),
+              backgroundColor: Colors.red,
+              duration: Duration(seconds: 2),
+            ),
+          );
+        });
+      }
     }
   }
 }