+++ lib/src/ui/reporter_screen.dart
@@ -3,20 +3,13 @@ import 'package:log_reporter/app_log_reporter.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 
 class AppLogReporterScreen extends StatelessWidget {
-  const AppLogReporterScreen({super.key, required this.reporter});
-  final SmartEmailReporter reporter;
+  const AppLogReporterScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
     return TalkerScreen(
       talker: AppLogReporter.talker,
       appBarTitle: AppLogReporter.config.appName,
-      appBarLeading: IconButton(
-        onPressed: () async {
-          await reporter.sendReport();
-        },
-        icon: Icon(Icons.send_outlined),
-      ),
     );
   }
 }